feat: enhance action mechanics and localization for DevourPeople, HelpPeople, and PlunderPeople
- Updated the DevourPeople action to utilize a new population loss ratio and introduced a mechanism for tracking devoured souls with a maximum limit.
- Revised the HelpPeople action to adjust population gain and cost, enhancing its mechanics and user experience.
- Improved the PlunderPeople action to reflect changes in population loss and gain, ensuring consistency with the new mechanics.
- Added new persistent effects for each action to influence luck based on their outcomes.
- Enhanced localization files to reflect changes in action descriptions and effects, improving clarity and cultural relevance across multiple languages.
- Updated tests to validate the new mechanics and ensure proper functionality of the actions in city regions.

+++ src/classes/core/avatar/core.py
@@ -104,7 +104,7 @@ class Avatar(
     backstory: Optional[str] = None
     emotion: EmotionType = EmotionType.CALM
     custom_pic_id: Optional[int] = None
-    luck_base: int = 0
+    luck_base: float = 0.0
     
     elixirs: List[ConsumedElixir] = field(default_factory=list)
     # 临时效果列表: [{"source": str, "effects": dict, "start_month": int, "duration": int}]
